Abstract AbstractDelayedDeliveryTracker & remove `super.getPriorityQueue()` & rebase master
